Filter redux store init with URL instead of component mount
