+++ packages/venia-concept/src/components/FilterModal/filterModal.js
@@ -29,45 +29,6 @@ class FilterModal extends Component {
         closeModalHandler: PropTypes.func
     };
 
-    getFilterParams = location => {
-        const params = new URLSearchParams(location.search);
-        let titles,
-            values = [];
-
-        let urlFilterParams = {};
-
-        for (var key of params.keys()) {
-            const cleanKey = key.replace(/\[\]\[.*\]/gm, '');
-
-            if (urlFilterParams[cleanKey]) continue;
-
-            titles = params.getAll(`${cleanKey}[][title]`);
-            values = params.getAll(`${cleanKey}[][value]`);
-
-            urlFilterParams[cleanKey] = titles.map((title, index) => ({
-                title: title,
-                value: values[index]
-            }));
-        }
-
-        return urlFilterParams;
-    };
-
-    componentDidMount = () => {
-        const filterParams = this.getFilterParams(this.props.history.location);
-        for (var key in filterParams) {
-            if (filterParams.hasOwnProperty(key)) {
-                filterParams[key].map(({ title, value }) => {
-                    this.props.filterAdd({
-                        group: key,
-                        title,
-                        value
-                    });
-                });
-            }
-        }
-    };
-
     render() {
         const { classes, isModalOpen, closeModalHandler } = this.props;
         const modalClass = isModalOpen ? classes.rootOpen : classes.root;
