+++ packages/venia-concept/src/components/FilterModal/filterModalContainer.js
@@ -2,7 +2,6 @@ import { connect } from 'react-redux';
 import { compose } from 'redux';
 import FilterModal from './filterModal';
 import { filterAdd, filterRemove } from 'src/actions/catalog';
-import { withRouter } from 'react-router-dom';
 
 const mapDispatchToProps = {
     filterAdd,
@@ -10,7 +9,6 @@ const mapDispatchToProps = {
 };
 
 export default compose(
-    withRouter,
     connect(
         null,
         mapDispatchToProps
